Alter clearOutputDir so we don't delete hidden files
git-svn-id: https://igv.googlecode.com/svn/trunk@2145 17392f64-ead8-4cea-ae29-09b3ab513800

+++ test/src/org/broad/igv/util/TestUtils.java
@@ -195,7 +195,10 @@ public class TestUtils {
         if (outputDir.isDirectory()) {
             File[] listFiles = outputDir.listFiles();
             for (File fi : listFiles) {
-                fi.delete();
+                //Keep hidden files and directories
+                if (!fi.getName().startsWith(".")) {
+                    fi.delete();
+                }
             }
         }
     }
